+++ apps/router-demo/router-remote2-2002/src/button.tsx
@@ -1,0 +1,3 @@
+export default function Button() {
+  return <div style={{ color: 'red' }}>Provider button</div>;
+}