+++ packages/bridge/bridge-react/src/index.ts
@@ -1,0 +1,6 @@
+export { createRemoteComponent } from './create';
+export { createBridgeComponent } from './provider';
+export type {
+  ProviderParams,
+  RenderFnParams,
+} from '@module-federation/bridge-shared';