+++ packages/bridge/vue3-bridge/src/index.ts
@@ -1,0 +1,3 @@
+export { createBridgeComponent } from './provider';
+export { createRemoteComponent } from './create';
+export type { RenderFnParams } from '@module-federation/bridge-shared';