+++ packages/enhanced/src/lib/container/ModuleFederationPlugin.ts
@@ -19,6 +19,7 @@ import SharePlugin from '../sharing/SharePlugin';
 import ContainerPlugin from './ContainerPlugin';
 import ContainerReferencePlugin from './ContainerReferencePlugin';
 import FederationRuntimePlugin from './runtime/FederationRuntimePlugin';
+import { RemoteEntryPlugin } from './runtime/RemoteEntryPlugin';
 
 const isValidExternalsType = require(
   normalizeWebpackPath(
@@ -70,6 +71,12 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
    */
   apply(compiler: Compiler): void {
     const { _options: options } = this;
+    // must before ModuleFederationPlugin
+    if (options.getPublicPath && options.name) {
+      new RemoteEntryPlugin(options.name, options.getPublicPath).apply(
+        compiler,
+      );
+    }
     if (options.dts !== false) {
       new DtsPlugin(options).apply(compiler);
     }