+++ packages/runtime/src/core.ts
@@ -94,6 +94,7 @@ export class FederationHost {
       [
         {
           url: string;
+          attrs?: Record<string, any>;
         },
       ],
       CreateScriptHookReturn
@@ -110,7 +111,7 @@ export class FederationHost {
     fetch: new AsyncHook<
       [string, RequestInit],
       Promise<Response> | void | false
-    >('fetch'),
+    >(),
   });
 
   constructor(userOptions: UserOptions) {