+++ packages/runtime/src/module/index.ts
@@ -34,8 +34,11 @@ class Module {
     const remoteEntryExports = await getRemoteEntry({
       remoteInfo: this.remoteInfo,
       remoteEntryExports: this.remoteEntryExports,
-      createScriptHook: (url: string) => {
-        const res = this.host.loaderHook.lifecycle.createScript.emit({ url });
+      createScriptHook: (url: string, attrs: any) => {
+        const res = this.host.loaderHook.lifecycle.createScript.emit({
+          url,
+          attrs,
+        });
 
         if (!res) return;