+++ packages/webpack-bundler-runtime/src/constant.ts
@@ -1,2 +1,2 @@
-export const ENCODE_NAME_PREFIX = 'ENCODE_NAME_PREFIX';
 export const FEDERATION_SUPPORTED_TYPES = ['script'];
+export { ENCODE_NAME_PREFIX } from '@module-federation/sdk';