+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,64 @@
 # [8.1.0-canary.7](https://github.com/module-federation/core/compare/nextjs-mf-8.1.0-canary.6...nextjs-mf-8.1.0-canary.7) (2023-11-21)
 
+## 8.3.26
+
+### Patch Changes
+
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+  - @module-federation/sdk@0.2.3
+  - @module-federation/node@2.3.0
+  - @module-federation/runtime@0.2.3
+  - @module-federation/enhanced@0.2.3
+  - @module-federation/utilities@3.0.30
+
+## 8.3.25
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.2
+- @module-federation/node@2.2.15
+- @module-federation/runtime@0.2.2
+- @module-federation/sdk@0.2.2
+- @module-federation/utilities@3.0.29
+
+## 8.3.24
+
+### Patch Changes
+
+- a047837: async function enable to stop warnings
+- Updated dependencies [88445e7]
+  - @module-federation/enhanced@0.2.1
+  - @module-federation/sdk@0.2.1
+  - @module-federation/node@2.2.14
+  - @module-federation/runtime@0.2.1
+  - @module-federation/utilities@3.0.28
+
+## 8.3.23
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.0
+- @module-federation/node@2.2.13
+- @module-federation/runtime@0.2.0
+- @module-federation/sdk@0.2.0
+- @module-federation/utilities@3.0.27
+
+## 8.3.22
+
+### Patch Changes
+
+- Updated dependencies [88900ad]
+- Updated dependencies [1e93c5e]
+  - @module-federation/sdk@0.1.21
+  - @module-federation/enhanced@0.1.21
+  - @module-federation/node@2.2.12
+  - @module-federation/runtime@0.1.21
+  - @module-federation/utilities@3.0.26
+
 ## 8.3.21
 
 ### Patch Changes