+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @module-federation/runtime
 
+## 0.2.3
+
+### Patch Changes
+
+- 32f26af: fix fetch hook types on runtime plugin interfaces
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+  - @module-federation/sdk@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [88445e7]
+  - @module-federation/sdk@0.2.1
+
+## 0.2.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.0
+
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [88900ad]
+  - @module-federation/sdk@0.1.21
+
 ## 0.1.20
 
 ### Patch Changes