+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.2.3
+
+## 0.2.2
+
+## 0.2.1
+
+## 0.2.0
+
+## 0.1.21
+
 ## 0.1.20
 
 ### Patch Changes